+++ ChangeLog.md
@@ -4,6 +4,12 @@
 
 - NEW: Updates the travel amount when selecting the destination country - *2024-03-20* 1.6
 
+## Release 1.5
+
+- NEW: Create a new module conf to set a default amount for travel packages - *2024-03-19* 1.5
+  - Create a new dictionary to define a default amount for each country
+  - Set new travel package amount with a default one if no one is given
+
 ## Release 1.4
 
 - NEW: Create box to display latest travel packages - *2024-03-19* 1.4